Fix backward search bug: no backing wall
We were allowing the backward search to happen in a direction that
didn't have a backing wall. This made it seem that we could stop
anywhere without hitting a wall and led to spurious solutions.

Now a backward search is only started if it is in the direction opposite
to a wall.